create alias now appropriately checks for errors.
