+++ mail-aliases/src/app/components/inputField.js
@@ -1,40 +1,12 @@
-import { Button } from 'react-bootstrap';
-import { spaceSlug, adjective, noun } from 'space-slug';
+import Form from 'react-bootstrap/Form';
 
-export function RandomizerButton() {
-  function randomizer() {
-    const createInputField = document.getElementsByName('create')[0];
-    const queryText = createInputField.value;
-
-    let newString = '';
-    if (queryText.indexOf('.') > 0) {
-      // Only keep the first part
-      newString = queryText.split('.')[0] + '.' + generateRandomString(4);
-    } else {
-      newString = spaceSlug([ adjective(1), noun(1) ], { separator: '' });
-    }
-
-    createInputField.value = newString;
-  }
-
+export function QueryInputField({ fieldName, defaultValue, className = '' }) {
   return (
-    <Button variant='outline-secondary' id='randomizer' onClick={randomizer}>
-      <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" className="bi bi-shuffle" viewBox="0 0 16 16">
-        <path fillRule="evenodd" d="M0 3.5A.5.5 0 0 1 .5 3H1c2.202 0 3.827 1.24 4.874 2.418.49.552.865 1.102 1.126 1.532.26-.43.636-.98 1.126-1.532C9.173 4.24 10.798 3 13 3v1c-1.798 0-3.173 1.01-4.126 2.082A9.624 9.624 0 0 0 7.556 8a9.624 9.624 0 0 0 1.317 1.918C9.828 10.99 11.204 12 13 12v1c-2.202 0-3.827-1.24-4.874-2.418A10.595 10.595 0 0 1 7 9.05c-.26.43-.636.98-1.126 1.532C4.827 11.76 3.202 13 1 13H.5a.5.5 0 0 1 0-1H1c1.798 0 3.173-1.01 4.126-2.082A9.624 9.624 0 0 0 6.444 8a9.624 9.624 0 0 0-1.317-1.918C4.172 5.01 2.796 4 1 4H.5a.5.5 0 0 1-.5-.5"/>
-        <path d="M13 5.466V1.534a.25.25 0 0 1 .41-.192l2.36 1.966c.12.1.12.284 0 .384l-2.36 1.966a.25.25 0 0 1-.41-.192zm0 9v-3.932a.25.25 0 0 1 .41-.192l2.36 1.966c.12.1.12.284 0 .384l-2.36 1.966a.25.25 0 0 1-.41-.192z"/>
-      </svg>
-    </Button>
+    <>
+      <Form.Control type="text" name={fieldName} defaultValue={defaultValue} className={className} required="" />
+      <Form.Control.Feedback type="invalid">
+        Please provide input in this field.
+      </Form.Control.Feedback>
+    </>
   );
-}
-
-function generateRandomString(string_length = 4) {
-  const chars = '0123456789abcdefghiklmnopqrstuvwxyz';
-  let randomstring = '';
-
-  for (let i = 0; i < string_length; i++) {
-    const rnum = Math.floor(Math.random() * chars.length);
-    randomstring += chars.substring(rnum, rnum + 1);
-  }
-
-  return randomstring;
 }
